Fixed bug where chat client would crash mobile app.
Implemented live-updating ETA for the Android mobile app along with custom
infowindows for the map. Right now, though, it goes off of the stop
closest to the shuttle, so we to make a list of stops in the order the
shuttle goes, so we can check one off and not have it try to estimate the
time after it's already went by.

** Bug** The mobile app now handles not having access to the network, it prompts
the user with a "no network" dialogue, but after the app exits, it crashes.

Markers now update based on their shuttle's name instead of the info.
That's why they weren't actually clearing off the map, haha. Got that
taken care of though.

+++ TrackerApps/TrackerTransmitter/src/com/cwardcode/TranTracker/LocReceiver.java
@@ -51,7 +51,7 @@ public class LocReceiver extends BroadcastReceiver {
 		Double longitude = intent.getDoubleExtra("longitude", -1);
 		Double speed = intent.getDoubleExtra("speed", -1);
 		String nextStop = intent.getStringExtra("nextStop");
-		Double estWait = intent.getDoubleExtra("estWait", -133);
+		int estWait = intent.getIntExtra("estWait", -133);
 		updateRemote(vehicleName, latitude, longitude, speed, nextStop, estWait);
 	}
